+++ src/main/kotlin/com/atlassian/performance/tools/report/api/result/DurationData.kt
@@ -3,6 +3,9 @@ package com.atlassian.performance.tools.report.api.result
 import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics
 import java.time.Duration
 
+/**
+ * @deprecated The generated `copy` and `componentN` methods should not be used. It will become a non-data class.
+ */
 data class DurationData(
     val stats: DescriptiveStatistics,
     internal val durationMapping: (Double) -> Duration